Fix crash due to ratelimiter dependency
Fixes: https://github.com/grokability/UnifiSnipeSync/issues/1

The rate limiting library in use has been long abandoned, and it doesn't
work with newer pythons because of changes to asyncio in 3.11.

It did include a synchronous version of the rate limiter, though, which is fine for
our needs.

There are better rate limiting libraries out there in terms of type of
rate limiter, but I didn't want to adapt them into this code, as most of
the better ones (eg: rush) are very specifically written for limiting
incoming requests, not outgoing ones. Might not be a big deal to do, but
this was even easier. If folks really want better rate limiting in this
tool, we can look into it when the tool is actually functional again.